requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
boto3==1.17.8
botocore==1.20.8
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-environ==0.4.5
django-json-widget==1.1.0
django-storages==1.11.1
djangorestframework==3.12.2
drf-nested-routers==0.92.5
future==0.18.2
google-auth==1.25.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httpie==2.4.0
idna==2.10
jmespath==0.10.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
